Add PlotsExt and first 1D recipe
