+++ Project.toml
@@ -6,19 +6,24 @@ version = "0.0.2"
 [deps]
 Debugger = "31a5f54b-26ea-5ae9-a837-f05ce5417438"
 DimensionalData = "0703355e-b756-11e9-17c0-8b28908087d0"
-RecipesBase = "3cdcf5f2-1ef4-517c-9805-6587b60abb01"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 SafeTestsets = "1bc83da4-3b8d-516f-aca4-4fe02f6d838f"
-Unitful = "1986cc42-f94f-5a68-af5c-568840ba703d"
+
+[weakdeps]
+Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
+
+[extensions]
+PlotsExt = "Plots"
 
 [compat]
+julia = "~1.9"
 DimensionalData = "~0.24"
 Reexport = "0.2,1"
-julia = "~1.8"
+Plots = "~1.38"
 
 [extras]
 SafeTestsets = "1bc83da4-3b8d-516f-aca4-4fe02f6d838f"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["Test", "SafeTestsets"]
+test = ["Test", "SafeTestsets", "Plots"]
