--- src/NMRPlots/NMRPlots.jl
@@ -1,8 +1,0 @@
-module NMRPlots
-
-using ..NMRBase
-
-using RecipesBase: @recipe, @series
-
-
-end
